+++ src/Model/Errors.js
@@ -3,10 +3,10 @@ import { without } from "../utils"
 export default class Errors {
   constructor({ _schema, ...errors }) {
     this._schema = without.call(_schema, "_primaryKey", "_timestamps")
-    Object.keys(_schema).map(attr => (this[attr] = errors[attr] || []))
+    Object.keys(_schema).forEach(attr => (this[attr] = errors[attr] || []))
   }
   clear() {
-    Object.keys(this._schema).map(attr => (this[attr] = []))
+    Object.keys(this._schema).forEach(attr => (this[attr] = []))
   }
 
   get any() {